Jolt/requirements.txt
bs4 == 0.0.1
gsearch == 1.6.0
newspaper3k == 0.2.5
python_telegram_bot == 12.0.0b1
